chore: restructure upload modules
