+++ config/service_registration.py
@@ -12,8 +12,8 @@ from services.upload.protocols import (
 def register_upload_services(container: ServiceContainer) -> None:
     """Register upload-related services with the container."""
 
-    from services.upload.processing import UploadProcessingService
-    from services.upload.validators import ClientSideValidator
+    from services.upload.core.processor import UploadProcessingService
+    from services.upload.core.validator import ClientSideValidator
     from services.data_processing.async_file_processor import AsyncFileProcessor
     from utils.upload_store import UploadedDataStore
